перенес typings в корень проекта, исправил все подсказки от pyright: typeCheckingMode - basic

+++ ugc-service/api/v1/models/ugc.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel, Field
 class RatingModel(BaseModel):
     user_id: UUID = Field(description="Идентификатор юзера", title="Идентификатор")
     film_id: UUID = Field(description="Идентификатор фильма", title="Идентификатор")
-    rating: int = Field(description="Рейтинг фильма", title="Рейтинг")
+    rating: float = Field(description="Рейтинг фильма", title="Рейтинг")
 
 
 class ReviewModel(BaseModel):
